+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Thumbnail asset on Collection and Item.
 - Metadata Asset on STAC Collection.
 
 ### Deprecated